+++ server/app/controllers/book/book_controller.py
@@ -1,53 +1,42 @@
 from flask import request, jsonify
-from . import user_api
-from ...services.user.user_service import(
-    list_users, get_user_byId, update_user_byId, delete_user_byId
 
+from . import book_api
+from ...services.book.book_service import (
+    list_books, get_book_by_id, save_new_book, update_book_info,
+    delete_book_from_db, search_books_by_query
 )
 from ...utils.decorators import JWT_required, admin_required
 
 
-@user_api.route("/", method = ["get"])
+@book_api.route("/")
 @JWT_required
-@admin_required
-def get_all_users():
-    """
-    API để lấy danh sách tất cả người dùng.
-    """
+def get_all_books():
     try:
-    
-        users = list_users()
+        books_data = list_books()
         return jsonify({
             "success": True,
-            "message": "Successfully fetched all users.",
-            "users": [user.as_dict() for user in users]
+            "message": "Successfully fetched all books.",
+            "books": books_data
         }), 200
+    
     except Exception as e:
         return jsonify({
             "error": "Internal server error.",
             "message": str(e)
         }), 500
-    
-@user_api.route("<user_id>", method = ["GET "])
+
+
+@book_api.route("/<book_id>")
 @JWT_required
-@admin_required
-def get_user(user_id):
-    """
-    API để lấy thông tin chi tiết của một người dùng dựa trên ID.
-    """
+def get_book(book_id):
     try:
-        # Gọi hàm get_user_byId từ service
-        user = get_user_byId(user_id)
-        if not user:
-            return jsonify({
-                "success": False,
-                "message": "User not found."
-            }), 404
+        book_data = get_book_by_id(book_id)
         return jsonify({
             "success": True,
-            "message": "Successfully fetched user.",
-            "user": user.as_dict()
+            "message": "Successfully fetched book.",
+            "book": book_data
         }), 200
+    
     except Exception as e:
         return jsonify({
             "error": "Internal server error.",
@@ -55,61 +44,129 @@ def get_user(user_id):
         }), 500
         
         
-@user_api.route("/<user_id>", methods=["PUT"])
+@book_api.route("/", methods=["POST"])
 @JWT_required
 @admin_required
-def update_user(user_id):
-    """
-    API để cập nhật thông tin người dùng dựa trên ID.
-    """
+def add_book():
     try:
-        
         data = request.get_json()
         if data is None:
             raise ValueError("Invalid JSON data.")
-
-      
-        updated_user = update_user_byId(user_id, data)
-        if not updated_user:
+        
+        REQUIRED_FIELDS = {"title", "author", "image_url", "description", "price", "quantity"}
+        missing_fields = {field for field in REQUIRED_FIELDS if data.get(field) is None}
+        
+        if missing_fields:
             return jsonify({
                 "success": False,
-                "message": "User not found or no data updated."
-            }), 404
+                "message": f"Missing required fields: {', '.join(missing_fields)}"
+            }), 400
+            
+        new_book = save_new_book(data["title"], data["author"], data["image_url"], data["description"], data["price"], data["quantity"])
         return jsonify({
             "success": True,
-            "message": "Successfully updated user.",
-            "user": updated_user.as_dict()
-        }), 200
+            "message": "Successfully added new book.",
+            "new_book": new_book.as_dict()
+        }), 201
+        
     except ValueError as e:
         return jsonify({
             "error": "Invalid data.",
             "message": str(e)
         }), 400
+    
     except Exception as e:
         return jsonify({
             "error": "Internal server error.",
             "message": str(e)
         }), 500
 
-@user_api.route("/<user_id>", methods=["DELETE"])
+
+@book_api.route("/<book_id>", methods=["PUT"])
 @JWT_required
 @admin_required
-def delete_user(user_id):
-    """
-    API để xóa một người dùng dựa trên ID.
-    """
+def update_book(book_id):
     try:
-        
-        deleted = delete_user_byId(user_id)
-        if not deleted:
+        book = get_book_by_id(book_id)
+        if not book:
             return jsonify({
                 "success": False,
-                "message": "User not found."
+                "message": "Book not found."
             }), 404
+            
+        data = request.get_json()
+        if data is None:
+            raise ValueError("Invalid JSON data.")
+        
+        ALLOW_FIELDS = {"title", "author", "image_url", "description", "price", "quantity"}
+        unknown_fields = {field for field in data if field not in ALLOW_FIELDS}
+        if unknown_fields:
+            return jsonify({
+                "success": False,
+                "message": f"Unknown fields: {', '.join(unknown_fields)}"
+            }), 400
+            
+        updated_book = update_book_info(book, data)
         return jsonify({
             "success": True,
-            "message": "User deleted successfully."
+            "message": "Successfully updated book.",
+            "updated_book": updated_book.as_dict()
         }), 200
+        
+    except ValueError as e:
+        return jsonify({
+            "error": "Invalid data.",
+            "message": str(e)
+        }), 400
+    
+    except Exception as e:
+        return jsonify({
+            "error": "Internal server error.",
+            "message": str(e)
+        }), 500
+        
+        
+@book_api.route("/<book_id>", methods=["DELETE"])
+@JWT_required
+@admin_required
+def delete_book(book_id):
+    try:
+        book = get_book_by_id(book_id)
+        if not book:
+            return jsonify({
+                "success": False,
+                "message": "Book not found."
+            }), 404
+            
+        delete_book_from_db(book)
+        return "", 204
+        
+    except Exception as e:
+        return jsonify({
+            "error": "Internal server error.",
+            "message": str(e)
+        }), 500
+        
+        
+@book_api.route("/search", methods=["GET"])
+@JWT_required
+def search_books():
+    try:
+        query = request.args.get("query", type=str, default=None)
+        if not query:
+            return jsonify({
+                "success": False,
+                "message": "Missing query parameter."
+            }), 400
+        
+        book_search_results = search_books_by_query(query)
+        return jsonify({
+            "success": True,
+            "message": "Search completed successfully.",
+            "total": len(book_search_results),
+            "books": book_search_results
+        }), 200
+    
     except Exception as e:
         return jsonify({
             "error": "Internal server error.",
